+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## [Unreleased][unreleased]
 
+## Added
+- `context.getProviderFromEmail(email:string)`
+
 ## Changed
 - rename `DeltaChat` to `AccountManager`
 - rename `DeltaChat.accounts()` to `AccountManager.getAllAccountIds()`
@@ -9,6 +12,9 @@
 ## Fixed
 - Fix documentation comment of `AccountManager`
 - Fix configure for multiple accounts at once
+
+## Deprecated
+- `DeltaChat.getProviderFromEmail(email:string)`, please use `context.getProviderFromEmail(email:string)` instead
 
 ## [1.60.0] - 2021-01-09